requirements.txt
langgraph==0.2.28
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.7
faiss-cpu==1.9.0
fastapi==0.115.4
uvicorn==0.32.0
streamlit==1.40.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
sentence-transformers==3.2.1
tiktoken==0.8.0
# Benchmarking & Evaluation
langfuse==2.36.1
ragas==0.1.16
datasets==2.19.1
